app/requirements.txt
databricks-sdk
pandas
numpy
pyarrow
databricks-sql-connector
plotly
streamlit
h3
keplergl
streamlit-keplergl
pytz
